my-native-book-classic/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb>=0.5.0
openai>=1.13.3
python-dotenv==1.0.0
PyMuPDF==1.23.8
markdownify==0.11.6
langchain==0.0.339
tiktoken>=0.7.0
pypdf==3.17.1